feature: Adicionado realizacao de emprestimos

+++ src/main/java/model/Emprestimo.java
@@ -2,7 +2,7 @@ package model;
 
 import java.time.LocalDate;
 
-public record Emprestimo(Integer id_cliente,
-                         Integer id_livro,
+public record Emprestimo(Long id_cliente,
+                         Long id_livro,
                          LocalDate data_emprestimo) {
 }
